WIP: implementors [build-system] Make the file build settings query asynchronous
The end goal is that build settings queries will register a delegate
that can asynchronously receive the settings, and be called when they
change. As a step towards that, start by making the API asynchronous
with a completion block. We may want to keep that as a convenience API
anyway for clients that don't care about changes to the settings.